// src/helpers/types/ClientRezultType.ts
export interface ClientRezType   {
    "id": number;
    "firstName": string;
    "lastName": string
    "userId": null;
    "categoryName": null;
    "correctAnswers": number;
    "countAnswers": number;
    "extraResDtoList": [] | null;
    "durationTime": number;
    "createdAt": string ;
    "status": string;
    "testScore": number;
    "fileId": null;
  }